add notification, fixed style css
