+++ src/js/handlers/onPlayBtnClick.js
@@ -1,6 +1,9 @@
 export { onBtnPlayClick };
-import searchTrailerById from '../searchTailerById';
+import ApiService from '../api-service';
 import traillerMarkup from '../templates/traillerMarkup';
+
+const api = new ApiService();
+
 async function onBtnPlayClick(e) {
   const backdropEl = document.querySelector('.backdrop');
   const modalEl = document.querySelector('.modal');
@@ -9,13 +12,18 @@ async function onBtnPlayClick(e) {
   const closeTrailerlBtn = document.querySelector('.modal__btn-close-trailer');
   if (e.target.classList.contains('modal__btn-play')) {
     const filmId = e.currentTarget.getAttribute('id');
-    const key = await searchTrailerById(filmId);
-    const markup = await traillerMarkup(key);
-    modalCardEl.classList.add('is-hidden');
-    closeModalBtn.classList.add('is-hidden');
-    closeTrailerlBtn.classList.remove('is-hidden');
-    modalEl.insertAdjacentHTML('afterbegin', markup);
-    backdropEl.addEventListener('click', onBtnCloseTrailer);
+    try {
+      const key = await api.searchTrailerById(filmId);
+      const markup = await traillerMarkup(key);
+      modalCardEl.classList.add('is-hidden');
+      closeModalBtn.classList.add('is-hidden');
+      closeTrailerlBtn.classList.remove('is-hidden');
+      modalEl.insertAdjacentHTML('afterbegin', markup);
+      backdropEl.addEventListener('click', onBtnCloseTrailer);
+    } catch (error) {
+      swal('sorry, there are no trailers for this movie(');
+      console.log(error);
+    }
   }
 }
 function onBtnCloseTrailer(e) {
@@ -26,7 +34,6 @@ function onBtnCloseTrailer(e) {
   const closeTrailerlBtn = document.querySelector('.modal__btn-close-trailer');
   if (e.target.classList.contains('modal__btn-close-trailer')) {
     modalVideoEl.remove();
-    console.log('on btn player click');
     modalCardEl.classList.remove('is-hidden');
     closeModalBtn.classList.remove('is-hidden');
     closeTrailerlBtn.classList.add('is-hidden');
