+++ src/js/templates/modalCardMarkap.js
@@ -1,6 +1,5 @@
 import icon from '../../images/sprite.svg';
-
-const IMG_URL = 'https://image.tmdb.org/t/p/w500/';
+import defaultPoster from '../../images/no-poster.jpg';
 
 export default function createModalMarkup({
   poster_path,
@@ -13,6 +12,9 @@ export default function createModalMarkup({
   overview,
 }) {
   const genersList = genres.map(element => element.name);
+  const imageUrl = poster_path
+    ? `https://image.tmdb.org/t/p/w500/${poster_path}`
+    : `${defaultPoster}`;
   return `<div class="modal">
     <button class="modal__close" data-action="close-modal"><svg class="modal__close-icon" width="14" height="14">
         <use href="${icon}#icon-close"></use>
@@ -24,7 +26,7 @@ export default function createModalMarkup({
         <div class="modal__content">
      
     <button class="modal__btn-play">play</button>
-    <img class="modal__img" src="${IMG_URL}${poster_path}" alt="cinema" />
+    <img class="modal__img" src="${imageUrl}" alt="cinema" />
   </div>
